use int32_t more

+++ sperl_allocator_util.c
@@ -13,7 +13,7 @@ void SPerl_ALLOCATOR_UTIL_exit_with_malloc_failure() {
 
 void* SPerl_ALLOCATOR_UTIL_safe_malloc(size_t count, size_t size) {
 
-  if (count <= 0 || size <= 0 || count > INT64_MAX / size) {
+  if (count <= 0 || size <= 0 || count > SIZE_MAX / size) {
     SPerl_ALLOCATOR_UTIL_exit_with_malloc_failure();
   }
   
@@ -29,7 +29,7 @@ void* SPerl_ALLOCATOR_UTIL_safe_malloc(size_t count, size_t size) {
 }
 
 void* SPerl_ALLOCATOR_UTIL_safe_malloc_zero(size_t count, size_t size) {
-  if (count <= 0 || size <= 0 || count > INT64_MAX / size) {
+  if (count <= 0 || size <= 0 || count > SIZE_MAX / size) {
     SPerl_ALLOCATOR_UTIL_exit_with_malloc_failure();
   }
   
@@ -43,7 +43,7 @@ void* SPerl_ALLOCATOR_UTIL_safe_malloc_zero(size_t count, size_t size) {
 }
 
 void* SPerl_ALLOCATOR_UTIL_safe_realloc(void* ptr, size_t count, size_t size) {
-  if (count <= 0 || size <= 0 || count > INT64_MAX / size) {
+  if (count <= 0 || size <= 0 || count > SIZE_MAX / size) {
     SPerl_ALLOCATOR_UTIL_exit_with_malloc_failure();
   }
